% Auto-generated: do not edit by hand
\name{''LoadingWrapper}

\alias{''LoadingWrapper}

\title{LoadingWrapper component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''LoadingWrapper(children=NULL, id=NULL, className=NULL, ignores=NULL,
interval=NULL, is_loading=NULL, targetClassName=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{className}{Character. }

\item{ignores}{List of characters. }

\item{interval}{Numeric. }

\item{is_loading}{Logical. }

\item{targetClassName}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

